# openapi.yml
openapi: 3.0.0
info:
  title: budgetManagement
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/v1/transactions:
    get:
      tags:
        - default
      summary: GET
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - default
      summary: POST
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                memo: test 1
                account:
                  id: 1
                amount: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - default
      summary: DELETE
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/transactions/:
    get:
      tags:
        - default
      summary: GET AccountID
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...